Start working on polling

+++ src/network/messages.rs
@@ -1,13 +1,14 @@
 use rkyv::{Archive, Deserialize, Serialize};
 
-#[repr(C)]
+#[repr(C, i32)]
+#[sharpify::client_poll]
 #[derive(Archive, Deserialize, Serialize, Debug, PartialEq)]
 pub enum ServerMessage {
     Pong,
     ChatMessage(String, String),
     NewUser(String),
     UserLeft(String),
-    Status { user_count: usize, tick_diff: f32 },
+    Status { user_count: u32, tick_diff: f32 },
 }
 
 #[repr(C)]
